改进安装过程
若出现错误导致安装过程中断，可以不再重复下载文件。

+++ install.sh
@@ -15,18 +15,34 @@ if [ $(id -u) != "0" ]; then
 fi
 
 #安装前删除resources文件夹下的源文件
-read -p "Files in ~/.lanmp/resources must be deleted before installation, press Enter to continue or Ctrl+C to quit this script."
-rm -rf ~/.lanmp/resources/libmcrypt*
-rm -rf ~/.lanmp/resources/mcrypt*
-rm -rf ~/.lanmp/resources/mhash*
-rm -rf ~/.lanmp/resources/bison*
-rm -rf ~/.lanmp/resources/cmake*
-rm -rf ~/.lanmp/resources/pcre*
-rm -rf ~/.lanmp/resources/apr*
-rm -rf ~/.lanmp/resources/nginx*
-rm -rf ~/.lanmp/resources/httpd*
-rm -rf ~/.lanmp/resources/mariadb*
-rm -rf ~/.lanmp/resources/php*
+read -p "Would you want to delete all files in ~/.lanmp/resources? " deleteresources
+if [ "$deleteresources" = "Y" ] || [ "$deleteresources" = "y" ]; then
+  rm -rf ~/.lanmp/resources/libmcrypt*
+  rm -rf ~/.lanmp/resources/mcrypt*
+  rm -rf ~/.lanmp/resources/mhash*
+  rm -rf ~/.lanmp/resources/bison*
+  rm -rf ~/.lanmp/resources/cmake*
+  rm -rf ~/.lanmp/resources/pcre*
+  rm -rf ~/.lanmp/resources/apr*
+  rm -rf ~/.lanmp/resources/nginx*
+  rm -rf ~/.lanmp/resources/httpd*
+  rm -rf ~/.lanmp/resources/mariadb*
+  rm -rf ~/.lanmp/resources/php*
+else
+  read -p ".tar.gz files in ~/.lanmp/resources will be used for Installing. Press Enter to continue, or Ctrl+C to exit the installation."
+  rm -rf ~/.lanmp/resources/libmcrypt-*
+  rm -rf ~/.lanmp/resources/mcrypt-*
+  rm -rf ~/.lanmp/resources/mhash-*
+  rm -rf ~/.lanmp/resources/bison-*
+  rm -rf ~/.lanmp/resources/cmake-*
+  rm -rf ~/.lanmp/resources/pcre-*
+  rm -rf ~/.lanmp/resources/apr-*
+  rm -rf ~/.lanmp/resources/nginx-*
+  rm -rf ~/.lanmp/resources/httpd-*
+  rm -rf ~/.lanmp/resources/mariadb-*
+  rm -rf ~/.lanmp/resources/php-*
+fi
+
 
 #检查用户需要安装什么
 WhatInstall="Noinstall"
@@ -123,22 +139,28 @@ echo "Starting Installation..."
 #询问要安装的Nginx版本
 #版本号参照http://nginx.org/en/download.html
 Nginx_Version(){
-echo "The version of Nginx can be referenced to http://nginx.org/en/download.html"
-read -p "Enter the version you want to install(1.8.0 by deafult):" vernginx
-if [ "$vernginx" = "" ]; then
-	vernginx="1.8.0"
-fi
-read -p "Do you want to install Nginx $vernginx?[Y/N]" ifinstallnginx
-if [ "$ifinstallnginx" = "Y" ] || [ "$ifinstallnginx" = "y" ]; then
+#判断是否已经存在源文件
+if [ ! -f "~/.lanmp/resources/nginx.tar.gz" ] then
+
+  echo "The version of Nginx can be referenced to http://nginx.org/en/download.html"
+  read -p "Enter the version you want to install(1.8.0 by deafult):" vernginx
+  if [ "$vernginx" = "" ]; then
+	   vernginx="1.8.0"
+  fi
+
+  read -p "Do you want to install Nginx $vernginx?[Y/N]" ifinstallnginx
+  if [ "$ifinstallnginx" = "Y" ] || [ "$ifinstallnginx" = "y" ]; then
 	echo "Trying downloading the Nginx..."
 	#尝试获取自定义的Nginx版本，若不成功，则退出
 	wget -O ~/.lanmp/resources/nginx.tar.gz http://nginx.org/download/nginx-$vernginx.tar.gz
-	if [ $? -ne 0 ]; then
-		echo "Nginx $vernginx cannot be downloaded.Please check if you have enter a correct version."
-		exit 1
-	fi
-else
-  echo "Installation interrupted.Please try again."
+	  if [ $? -ne 0 ]; then
+		    echo "Nginx $vernginx cannot be downloaded.Please check if you have enter a correct version."
+		    exit 1
+	  fi
+  else
+    echo "Installation interrupted.Please try again."
+  fi
+
 fi
 }
 
@@ -155,6 +177,8 @@ echo "=================================================="
 
 #询问要安装的MariaDB版本
 MariaDB_Version(){
+#判断是否已经存在源文件
+if [ ! -f "~/.lanmp/resources/mariadbin.tar.gz" ] then
   #下载二进制安装包
   read -p "Is your system 32bit or 64bit?(Enter 32 or 64)" sysbit
   if [ "$sysbit" = "32" ]; then
@@ -162,9 +186,9 @@ MariaDB_Version(){
   	rpm -qa|grep glibc
   	read -p "Is the version of glibc showed above over 2.14+?[Y/N]" verglibc
   	if [ "$verglibc" = "Y" ] || [ "$verglibc" = "y" ]; then
-  		wget -O ~/.lanmp/resources/mariadb.tar.gz https://downloads.mariadb.org/f/mariadb-10.1.9/bintar-linux-glibc_214-x86/mariadb-10.1.9-linux-glibc_214-i686.tar.gz
+  		wget -O ~/.lanmp/resources/mariadbin.tar.gz https://downloads.mariadb.org/f/mariadb-10.1.9/bintar-linux-glibc_214-x86/mariadb-10.1.9-linux-glibc_214-i686.tar.gz
   	elif [ "$verglibc" = "N" ] || [ "$verglibc" = "n" ]; then
-  		wget -O ~/.lanmp/resources/mariadb.tar.gz https://downloads.mariadb.org/f/mariadb-10.1.9/bintar-linux-x86/mariadb-10.1.9-linux-i686.tar.gz
+  		wget -O ~/.lanmp/resources/mariadbin.tar.gz https://downloads.mariadb.org/f/mariadb-10.1.9/bintar-linux-x86/mariadb-10.1.9-linux-i686.tar.gz
   	else
   		echo "Please enter Y or N.Try to run this script again."
   		exit 1
@@ -174,9 +198,9 @@ MariaDB_Version(){
   	rpm -qa|grep glibc
   	read -p "Is the version of glibc showed above over 2.14+?[Y/N]" verglibc
   	if [ "$verglibc" = "Y" ] || [ "$verglibc" = "y" ]; then
-  		wget -O ~/.lanmp/resources/mariadb.tar.gz https://downloads.mariadb.org/f/mariadb-10.1.9/bintar-linux-glibc_214-x86_64/mariadb-10.1.9-linux-glibc_214-x86_64.tar.gz
+  		wget -O ~/.lanmp/resources/mariadbin.tar.gz https://downloads.mariadb.org/f/mariadb-10.1.9/bintar-linux-glibc_214-x86_64/mariadb-10.1.9-linux-glibc_214-x86_64.tar.gz
   	elif [ "$verglibc" = "N" ] || [ "$verglibc" = "n" ]; then
-  		wget -O ~/.lanmp/resources/mariadb.tar.gz https://downloads.mariadb.org/f/mariadb-10.1.9/bintar-linux-x86_64/mariadb-10.1.9-linux-x86_64.tar.gz
+  		wget -O ~/.lanmp/resources/mariadbin.tar.gz https://downloads.mariadb.org/f/mariadb-10.1.9/bintar-linux-x86_64/mariadb-10.1.9-linux-x86_64.tar.gz
   	else
   		echo "Please enter Y or N.Try to run this script again."
   		exit 1
@@ -185,6 +209,7 @@ MariaDB_Version(){
   	echo "Cannot detect your system's type.Please enter a legal value."
   	exit 1
   fi
+fi
 }
 
 #引用安装文件
