+++ parts/mariadb.sh
@@ -17,7 +17,12 @@ useradd -r -g mysql mysql
 
 #下载编译
 cd ~/.lanmp/resources
-wget -O mariadb.tar.gz https://downloads.mariadb.org/f/mariadb-10.1.9/source/mariadb-10.1.9.tar.gz
+
+#判断是否已经存在源文件
+if [ ! -f "~/.lanmp/resources/mariadb.tar.gz" ] then
+  wget -O mariadb.tar.gz https://downloads.mariadb.org/f/mariadb-10.1.9/source/mariadb-10.1.9.tar.gz
+fi
+
 tar -zxf mariadb.tar.gz
 cd mariadb*
 cmake .
